Forgotten refactoring improvements
